requirements.txt
faiss-cpu==1.13.1
langchain==1.2.0
langchain-community==0.4.1
langchain-core==1.2.5
langchain-openai==1.1.6
langchain-text-splitters==1.1.0
pypdf==6.5.0
PyPDF2==3.0.1
python-dotenv==1.2.1
python-docx==1.2.0
reportlab==4.0.7